requirements.txt
ib_insync==0.9.86
nest_asyncio>=1.6.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.0.0
httpx>=0.27.0
tzdata>=2024.1
